use BSP2 internally
Prep for BSP2 support. File lumps are converted to internal structures which have more bits.

+++ src/bsp/bsptypes.cpp
@@ -4,7 +4,7 @@
 
 BSPEDGE::BSPEDGE() {}
 
-BSPEDGE::BSPEDGE(uint16_t v1, uint16_t v2) { 
+BSPEDGE::BSPEDGE(uint32_t v1, uint32_t v2) { 
 	iVertex[0] = v1;
 	iVertex[1] = v2; 
 }
